Add cookie to dummy helpers (closes #235)
Add cookie dependency to project.

Use cookie library to parse the value of the cookie header (if present)
and then extract the value of the requested key (if present).

Returns defaultValue in the pattern of the other dummy helpers if the
cookie is not set/no cookie header was provided.