Insert duplicates after originals (closes #239)
When duplicating routes and environments, insert the duplicate after the
original in the list (rather than at the end of the list).

Adds an optional "afterUUID" string to each of the addX actions and
splice the clone into the right place in the list in the reducer if
afterUUID is specified.

Add test coverage on duplicating routes and environments that covers the
insertion index. This could be extended to cover faithful duplication
but that's outside the scope of this MR.

Ignore vim swp files.