Fix bug with case sensitive Content-Type comparison
Content-Type was extracted from route/environment in a case sensitive manner which lead to wrong content type being applied to the route if capitalized one was not found. This could result in a JSON parsing error.
Also, uses includes instead of === to compare headers as additional info may be present in header (charset, etc)