Remove api prefix when auto-mocking #333
Check the prefix of the active environment when creating the route from the
environment logs. If the prefix is present in the request path then remove it
when creating the route.

Modify the environment-logs test spec to add a prefix and verify that the auto
mock removes the prefix in the created route.