Fix headers tests failing
On Linux, it seems that using setValue on fields does not immediately unfocused the previous field, which keep the typeaheads opened, and prevent clicks on add header buttons. Web driver should report add button as not clickable, but instead, subsequent headers entered overwrite the previous ones. The click ensure that the previous field is unfocused