Fix menu sizes not being restored in desktop app
Rely on the window state to maximize before showing the BrowserWindow. This avoid jumping from a smaller size to a maximize state and reduce the size of the menus.
Do not save the settings when restoring, but only after a resize event or a user interaction.
Also improve performances by only saving after mouse up or resize event.

Closes #1167