Remove search filter directive as it is incompatible trackby and external changes (#1264)
Use of trackbyUuid and elementRef.textContent, made it impossible for the directive to detect changes when data was modified from "outside" (e.g. cloud sync). The result was that when something was filtered, the displayed result could change to new content inserted at the same spot and completely unrelated.

Replaced with regular observables