Always display the body in the logs tab in an editor
The editor offers acceptable performances to always display the full body and stop truncating the content.
When JSON Content-Type is detected, try to beatify the JSON.
Also display a warning next to the "Last log" link to inform if one of the bodies was invalid.
Adding an editor also makes the body searchable, therefore closing #137
Closes #368, closes #137