Fix CRUD responses tab switching
If use navigates to a tab not available in the CRUD default response (the CRUD operations response) on a secondary response, like "callbacks" or "Settings", and switch back to the default response, the tab stays focused even if it doesn't exists.
Also fix CSS grid classes not being correctly selected depending on the route type.
Closes #1552